acl/requirements.txt
grpcio==1.12.1
grpcio-tools==1.12.1
protobuf==3.6.0
psycopg2-binary==2.7.5
SQLAlchemy==1.2.8